requirements.txt
Flask
pyodbc
openai
python-dotenv
flask-sqlalchemy